Move config-related exceptions to config.exceptions submodule
This is to avoid circular dependency between top-level .exceptions,
.api and .config.

Now .config module is self-contained (apart from a dependency on
.package_info), and .exceptions > .api > .config.
