+++ dwave/cloud/config/__init__.py
@@ -159,5 +159,6 @@ Examples:
 """
 
 from .constants import *
+from .exceptions import *
 from .loaders import *
 from .models import *
