+++ releasenotes/notes/fix-mod-deps-exceptions-api-config-8b7106aaffc9d778.yaml
@@ -1,0 +1,23 @@
+---
+features:
+  - |
+    Configuration constants like default region and default API endpoints are
+    now available in a new module ``dwave.cloud.config.constants``.
+
+    Configuration-related exceptions are now available in a new module
+    ``dwave.cloud.config.exceptions``.
+
+    ``dwave.cloud.config`` module (with submodules) is now free from dependencies
+    on other submodules, and it's safe to be imported in any cloud-client context.
+
+deprecations:
+  - |
+    Use of ``dwave.cloud.api.constants`` to access default configuration
+    constants is deprecated in ``dwave-cloud-client==0.13.1`` in favor of
+    a new config module ``dwave.cloud.config.constants``, and will be disabled
+    in ``dwave-cloud-client==0.14.0``.
+
+fixes:
+  - |
+    Fix a circular import error.
+    See `#669 <https://github.com/dwavesystems/dwave-cloud-client/issues/669>`_.
