+++ tests/__init__.py
@@ -16,7 +16,7 @@ import os
 import warnings
 
 from dwave.cloud.config import load_config
-from dwave.cloud.exceptions import ConfigFileError
+from dwave.cloud.config.exceptions import ConfigFileError
 
 
 # try to load client config needed for live tests on SAPI web service
